YAAUM-23 Fix issues
