+++ build-logic/convention/src/main/kotlin/dev/yaaum/convention/config/ProjectConfig.kt
@@ -15,7 +15,7 @@ object ProjectConfig {
         const val COMPILE_SDK = TARGET_SDK
     }
 
-    val JAVA_VERSION = JavaVersion.VERSION_18
+    val JAVA_VERSION = JavaVersion.VERSION_17
 
-    const val KOTLIN_VERSION = "18"
+    const val KOTLIN_VERSION = "17"
 }
