+++ common/core/build.gradle.kts
@@ -1,0 +1,16 @@
+@file:Suppress("SpellCheckingInspection")
+
+plugins {
+    id("yaaum.convention.feature")
+    id("yaaum.convention.common.detekt")
+    id("yaaum.convention.common.ktlint")
+}
+
+android {
+    namespace = "dev.yaaum.common.core"
+}
+
+dependencies {
+    implementation(libs.kotlinx.coroutines.android)
+    implementation(libs.kotlinx.coroutines.core)
+}
