+++ common/core/src/main/kotlin/dev/yaaum/common/core/error/BaseError.kt
@@ -1,0 +1,3 @@
+package dev.yaaum.common.core.error
+
+interface BaseError
