+++ common/konsist/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    id("yaaum.convention.feature")
+    id("yaaum.convention.common.detekt")
+    id("yaaum.convention.common.ktlint")
+    id("yaaum.convention.test.kotlin")
+}
+
+android {
+    namespace = "dev.yaaum.common.core.konsist"
+}
+
+
+dependencies{
+    testImplementation(libs.konsist)
+
+    testImplementation(project(":data:repository:core"))
+    testImplementation(project(":data:repository:timeusage"))
+    testImplementation(project(":data:repository:timeusage:impl"))
+
+    testImplementation(project(":domain:core"))
+    testImplementation(project(":domain:timeusage"))
+    testImplementation(project(":domain:timeusage:impl"))
+}
