--- core/common/build.gradle.kts
@@ -1,16 +1,0 @@
-@file:Suppress("SpellCheckingInspection")
-
-plugins {
-    id("yaaum.convention.feature")
-    id("yaaum.convention.common.detekt")
-    id("yaaum.convention.common.ktlint")
-}
-
-android {
-    namespace = "dev.yaaum.core.common"
-}
-
-dependencies {
-    implementation(libs.kotlinx.coroutines.android)
-    implementation(libs.kotlinx.coroutines.core)
-}
