--- core/common/src/main/kotlin/dev/yaaum/core/common/error/BaseError.kt
@@ -1,3 +1,0 @@
-package dev.yaaum.core.common.error
-
-interface BaseError
