--- presentation/core/common/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-    id("yaaum.convention.compose.feature")
-    id("yaaum.convention.common.detekt")
-    id("yaaum.convention.common.ktlint")
-}
-
-android {
-    namespace = "dev.yaaum.presentation.core.common"
-}
-
-dependencies{
-
-    implementation(project(":core:common"))
-}
