+++ presentation/core/platform/build.gradle.kts
@@ -1,13 +1,14 @@
 plugins {
-    id("yaaum.convention.feature")
+    id("yaaum.convention.compose.feature")
     id("yaaum.convention.common.detekt")
     id("yaaum.convention.common.ktlint")
 }
 
 android {
-    namespace = "dev.yaaum.domain.core"
+    namespace = "dev.yaaum.presentation.core.platform"
 }
 
-dependencies {
+dependencies{
+
     implementation(project(":common:core"))
 }
