--- presentation/feature/main/src/main/kotlin/dev/yaaum/main/navigation/MainScreenModule.kt
@@ -1,14 +1,0 @@
-package dev.yaaum.main.navigation
-
-import cafe.adriel.voyager.core.registry.screenModule
-import dev.yaaum.main.navigation.route.MainRoute
-import dev.yaaum.presentation.core.navigation.RouteGraph
-
-/**
- * Module for `presentation:feature:main`
- */
-val mainScreenModule = screenModule {
-    register<RouteGraph.MainScreen> { args ->
-        MainRoute(args.from)
-    }
-}
