--- presentation/feature/onboarding/src/main/kotlin/dev/yaaum/onboarding/navigation/OnboardingScreenModule.kt
@@ -1,14 +1,0 @@
-package dev.yaaum.onboarding.navigation
-
-import cafe.adriel.voyager.core.registry.screenModule
-import dev.yaaum.onboarding.navigation.route.OnboardingRoute
-import dev.yaaum.presentation.core.navigation.RouteGraph
-
-/**
- * Module for `presentation:feature:onboarding`
- */
-val onboardingScreenModule = screenModule {
-    register<RouteGraph.OnboardingScreen> {
-        OnboardingRoute()
-    }
-}
