--- presentation/feature/onboarding/src/main/kotlin/dev/yaaum/onboarding/navigation/route/OnboardingRoute.kt
@@ -1,20 +1,0 @@
-package dev.yaaum.onboarding.navigation.route
-
-import androidx.compose.runtime.Composable
-import cafe.adriel.voyager.core.screen.Screen
-import cafe.adriel.voyager.navigator.LocalNavigator
-import cafe.adriel.voyager.navigator.currentOrThrow
-import dev.yaaum.onboarding.screen.composable.onboarding.OnboardingScreen
-
-/**
- * Route for OnboardingScreen
- */
-class OnboardingRoute : Screen {
-    @Composable
-    override fun Content() {
-        val navigator = LocalNavigator.currentOrThrow
-//        val mainScreen = rememberScreen(SharedScreen.MainScreen)
-
-        OnboardingScreen(navigator)
-    }
-}
