+++ presentation/feature/settings/src/main/kotlin/dev/yaaum/presentation/feature/settings/navigation/route/AboutRoute.kt
@@ -1,0 +1,20 @@
+package dev.yaaum.presentation.feature.settings.navigation.route
+
+import androidx.compose.runtime.Composable
+import cafe.adriel.voyager.core.screen.Screen
+import cafe.adriel.voyager.navigator.LocalNavigator
+import cafe.adriel.voyager.navigator.currentOrThrow
+import dev.yaaum.presentation.feature.settings.screen.composable.about.AboutScreen
+
+/**
+ * Route for AboutScreen
+ */
+class AboutRoute : Screen {
+    @Composable
+    override fun Content() {
+        val navigator = LocalNavigator.currentOrThrow
+//        val mainScreen = rememberScreen(SharedScreen.MainScreen)
+
+        AboutScreen(navigator)
+    }
+}
